+++ modal-dialog/src/main/java/raven/modal/option/LayoutOption.java
@@ -35,6 +35,10 @@ public class LayoutOption {
         return animateDistance;
     }
 
+    public RelativeToOwnerType getRelativeToOwnerType() {
+        return relativeToOwnerType;
+    }
+
     public boolean isRelativeToOwner() {
         return relativeToOwner;
     }
@@ -47,11 +51,12 @@ public class LayoutOption {
         return onTop;
     }
 
-    private LayoutOption(DynamicSize location, Insets margin, DynamicSize size, DynamicSize animateDistance, boolean relativeToOwner, float animateScale, boolean onTop) {
+    private LayoutOption(DynamicSize location, Insets margin, DynamicSize size, DynamicSize animateDistance, RelativeToOwnerType relativeToOwnerType, boolean relativeToOwner, float animateScale, boolean onTop) {
         this.location = location;
         this.margin = margin;
         this.size = size;
         this.animateDistance = animateDistance;
+        this.relativeToOwnerType = relativeToOwnerType;
         this.relativeToOwner = relativeToOwner;
         this.animateScale = animateScale;
         this.onTop = onTop;
@@ -66,6 +71,7 @@ public class LayoutOption {
     private Insets margin = new Insets(7, 7, 7, 7);
     private DynamicSize size = new DynamicSize(-1, -1);
     private DynamicSize animateDistance = new DynamicSize(0, 20);
+    private RelativeToOwnerType relativeToOwnerType = RelativeToOwnerType.RELATIVE_CONTAINED;
     private boolean relativeToOwner;
     private float animateScale;
     private boolean onTop = false;
@@ -108,6 +114,11 @@ public class LayoutOption {
         return this;
     }
 
+    public LayoutOption setRelativeToOwnerType(RelativeToOwnerType relativeToOwnerType) {
+        this.relativeToOwnerType = relativeToOwnerType;
+        return this;
+    }
+
     public LayoutOption setRelativeToOwner(boolean relativeToOwner) {
         this.relativeToOwner = relativeToOwner;
         return this;
@@ -126,7 +137,16 @@ public class LayoutOption {
         return this;
     }
 
+    /**
+     * RELATIVE_CONTAINED: Modal and background are confined to the owner's bounds and track the owner's visibility. (default)
+     * RELATIVE_GLOBAL: Background spans the entire window.
+     * RELATIVE_BOUNDLESS: Background covers the owner, but the modal can extend outside the owner. Tracks owner's visibility. (requires heavyWeight = true)
+     */
+    public enum RelativeToOwnerType {
+        RELATIVE_CONTAINED, RELATIVE_GLOBAL, RELATIVE_BOUNDLESS
+    }
+
     public LayoutOption copy() {
-        return new LayoutOption(location, new Insets(margin.top, margin.left, margin.bottom, margin.right), new DynamicSize(size), new DynamicSize(animateDistance), relativeToOwner, animateScale, onTop);
+        return new LayoutOption(location, new Insets(margin.top, margin.left, margin.bottom, margin.right), new DynamicSize(size), new DynamicSize(animateDistance), relativeToOwnerType, relativeToOwner, animateScale, onTop);
     }
 }